Cleanup: line length

+++ release/scripts/startup/bl_operators/wm.py
@@ -2170,11 +2170,8 @@ class WM_OT_addon_remove(Operator):
         addon_utils.disable(self.module, default_set=True)
 
         import shutil
-        if isdir:
-            if os.path.islink(path):
-                os.remove(path)
-            else:
-                shutil.rmtree(path)
+        if isdir and (not os.path.islink(path)):
+            shutil.rmtree(path)
         else:
             os.remove(path)
